initial

+++ utils/general.py
@@ -368,7 +368,7 @@ def strip_optimizer(f='weights/best.pt', s=''):
     x['optimizer'] = None
     x['training_results'] = None
     x['epoch'] = -1
-    x['model'].half()  # to FP16
+    x['model'].float()  # to FP16
     for p in x['model'].parameters():
         p.requires_grad = False
     torch.save(x, s or f)
